Add a schema creation tool for gemonlinedb
The schemas are created from a definition file in JSON format. They use
common definitions found in gemonlinedb/schema/common.xsd. In the
future, the schema generation tool will be extended to also generate the
corresponding code.

The schemas' applicability is checked at build time by linting the example
files found in the gemonlinedb/xml folder. Since the tool used for this
validation is based on libxml2 and not Xerces, some incompatibilities might
be found later on.

See also issue #177.