Schema creation for gemonlinedb
* Add a schema creation tool for gemonlinedb
  * The schemas are created from a definition file in JSON format. They use
common definitions found in gemonlinedb/schema/common.xsd. In the
future, the schema generation tool will be extended to also generate the
corresponding code.
  * The schemas' applicability is checked at build time by linting the example
files found in the gemonlinedb/xml folder. Since the tool used for this
validation is based on libxml2 and not Xerces, some incompatibilities might
be found later on.
  * See also issue #177.
* Generate classes to hold database data
  *The generated classes will be used as base classes for the configuration
objects available in the public API. They live in the namespace
gem::onlinedb::detail and are abstract to prevent misuse.
  * See also issues #177 and #213.
* Integrate the Makefile into the build system
  * Rename test -> tests
  * Don't always run tests after building
  * Make tests depend on the default target
  * Clean generated headers
* There's still a problem with the default target: it tries to compile the
code before headers are generated, which fails if they are used. Then
compilation is performed a second time and succeeds. I have no idea why
this happens.
* Move gitignore rules to gemonlinedb/.gitignore