Create class XMLBuilder<Configuration>
The XML builder can be used to build XML files (represented using the
DOM) for configuration objects. It uses ConfigurationTraits to deduce
some of the values put into the XML files in a type-safe way.

There are still some problems to be understood with Xerces' memory
management. Passing DOM objects around as pointers without deleting
them should be mostly safe as they're owned by the implementation.
I'm more afraid of XMLChar * strings, as I believe that they are
owned by the user.