Add tests for XMLBuilder
The testing framework Boost.Test is used to simplify the code. Two tests
are added, both attempting to create an XML file with one run, one data
set and two data element with default-constructed VFAT3Configurations.

The XMLBuilderBasic test checks whether the DOM is being populated at all.
It fails early if XMLBuilder is very badly broken and detected mistakes
in early versions of its implementation. This test already passes.

The XMLBuilderXsdValidation attempts to perform XSD validation of the
generated XML. It was quite tricky to implement because of limitations in
Xerces. This test is currently failing because of two problems with the
XSD schema:

* Xerces doesn't allow multiple xs:include or xs:redefine tags
  referencing the same file. This will be fixed in parseDef.py.
* The schema enforces a specific order for its elements. This will be
  fixed in parseDef.py by using xs:all instead of xs:sequence.
* While enforced by the schema, the *_NAME-*_VALUE syntax isn't
  supported by XMLWriter. This limitation will remain until the
  situation of this syntax is clarified.

In addition, no attempt is made to find the xsd schema elsewhere than in
schema/ under the current directory.

See Github issues #177 and #215.