Turn XMLBuilder into XMLSerializationData
Realized that an XMLReader class turning DOM into configuration objects
would have the same contents as XMLBuilder, except that setters would be
turned into getters. Decided to merge them as XMLSerializationData.

In the future, the member function makeDOM could be changed to a free
function or a member of another class.