Add AMC13 configuration class
The AMC13 configuration includes a string (the hostname) and
RegisterData can only handle integer 32-bits values. Custom
serialization code is hence needed: in order to provide a consistent
interface, XMLSerializationData is specialized for AMC13Configuration.

Tests of the specializations are added to testAMC13Configuration.cc:
they duplicate testXMLSerialization.cc but make use of the new
specialization. This ensures that the specialization is consistent with
the original.